Move DeviceType and DeviceTypeMapping
